self-built/charts/postgres-operator/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
durationpy==0.9
frozenlist==1.5.0
google-auth==2.37.0
idna==3.10
iso8601==2.1.0
kopf==1.37.4
kubernetes==31.0.0
multidict==6.1.0
oauthlib==3.2.2
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
yarl==1.18.3
